#AG-27 updated sponsor for easier integration with events

+++ routes/web.php
@@ -16,14 +16,14 @@ Route::get('/', function () {
 
 Route::group(['prefix' => 'admin'], function () {
     Route::prefix('/sponsors')->group(function () {
-        Route::get('/', [SponsorController::class, 'index'])->name('sponsors.index');
+        Route::get('/', [SponsorController::class, 'index'])->name('admin.sponsors.index');
     });
     Route::prefix('/sponsor')->group(function () {
-        Route::get('/create', [SponsorController::class, 'create'])->name('sponsor.create');
-        Route::post('/store', [SponsorController::class, 'store'])->name('sponsor.store');
+        Route::get('/create', [SponsorController::class, 'create'])->name('admin.sponsor.create');
+        Route::post('/store', [SponsorController::class, 'store'])->name('admin.sponsor.store');
 
-        Route::get('/{id}', [SponsorController::class, 'sponsor'])->name('sponsor.show');
-        Route::put('/update/{id}', [SponsorController::class, 'update'])->name('sponsor.update');
-        Route::delete('/delete/{id}', [SponsorController::class, 'delete'])->name('sponsor.delete');
+        Route::get('/{id}', [SponsorController::class, 'show'])->name('admin.sponsor.show');
+        Route::put('/update/{id}', [SponsorController::class, 'update'])->name('admin.sponsor.update');
+        Route::delete('/delete/{id}', [SponsorController::class, 'delete'])->name('admin.sponsor.delete');
     });
-})->middleware('checkIfAdmin');
+});
